// src/types/PlayerStats.ts
export type PlayerStats = {
  id: number;
  playerId: number;
  sleeper_player_id: number;
  year: number;
  week: number;
  blk_kick?: number;
  bonus_fd_qb?: number;
  bonus_fd_rb?: number;
  bonus_fd_te?: number;
  bonus_fd_wr?: number;
  bonus_pass_cmp_25?: number;
  bonus_pass_yd_300?: number;
  bonus_rec_rb?: number;
  bonus_rec_te?: number;
  bonus_rec_wr?: number;
  bonus_rec_yd_100?: number;
  bonus_rush_att_20?: number;
  bonus_rush_rec_yd_100?: number;
  bonus_rush_yd_100?: number;
  bonus_sack_2p?: number;
  bonus_tkl_10p?: number;
  cmp_pct?: number;
  def_3_and_out?: number;
  def_4_and_stop?: number;
  def_forced_punts?: number;
  def_kr?: number;
  def_kr_lng?: number;
  def_kr_yd?: number;
  def_kr_ypa?: number;
  def_pass_def?: number;
  def_pr?: number;
  def_pr_lng?: number;
  def_pr_yd?: number;
  def_pr_ypa?: number;
  def_snp?: number;
  def_st_ff?: number;
  def_st_fum_rec?: number;
  def_st_tkl_solo?: number;
  def_td?: number;
  down_3_att?: number;
  down_3_conv?: number;
  down_3_pct?: number;
  down_4_att?: number;
  down_4_conv?: number;
  down_4_pct?: number;
  fan_pts_allow?: number;
  fan_pts_allow_k?: number;
  fan_pts_allow_qb?: number;
  fan_pts_allow_rb?: number;
  fan_pts_allow_te?: number;
  fan_pts_allow_wr?: number;
  fd?: number;
  ff?: number;
  fga?: number;
  fgm?: number;
  fgm_20_29?: number;
  fgm_30_39?: number;
  fgm_40_49?: number;
  fgm_50p?: number;
  fgm_lng?: number;
  fgm_pct?: number;
  fgm_yds?: number;
  fgm_yds_over_30?: number;
  fgmiss?: number;
  fgmiss_30_39?: number;
  fgmiss_40_49?: number;
  fgmiss_50p?: number;
  fum?: number;
  fum_lost?: number;
  fum_rec?: number;
  fum_ret_yd?: number;
  g2g_att?: number;
  g2g_conv?: number;
  g2g_pct?: number;
  gms_active?: number;
  gp?: number;
  gs?: number;
  idp_blk_kick?: number;
  idp_def_td?: number;
  idp_ff?: number;
  idp_fum_rec?: number;
  idp_fum_ret_yd?: number;
  idp_int?: number;
  idp_int_ret_yd?: number;
  idp_pass_def?: number;
  idp_pass_def_3p?: number;
  idp_qb_hit?: number;
  idp_sack?: number;
  idp_sack_yd?: number;
  idp_tkl?: number;
  idp_tkl_ast?: number;
  idp_tkl_loss?: number;
  idp_tkl_solo?: number;
  int?: number;
  int_ret_yd?: number;
  kick_pts?: number;
  kr?: number;
  kr_lng?: number;
  kr_yd?: number;
  kr_ypa?: number;
  off_snp?: number;
  off_yd?: number;
  off_yd_per_play?: number;
  opp_fd?: number;
  opp_off_yd?: number;
  opp_off_yd_per_play?: number;
  opp_pass_fd?: number;
  opp_rush_fd?: number;
  pass_air_yd?: number;
  pass_att?: number;
  pass_cmp?: number;
  pass_cmp_40p?: number;
  pass_fd?: number;
  pass_inc?: number;
  pass_int?: number;
  pass_int_td?: number;
  pass_lng?: number;
  pass_rtg?: number;
  pass_rush_yd?: number;
  pass_rz_att?: number;
  pass_sack?: number;
  pass_sack_yds?: number;
  pass_td?: number;
  pass_td_40p?: number;
  pass_td_50p?: number;
  pass_td_lng?: number;
  pass_yd?: number;
  pass_ypa?: number;
  pass_ypc?: number;
  penalty?: number;
  penalty_yd?: number;
  pos_rank_half_ppr?: number;
  pos_rank_ppr?: number;
  pos_rank_std?: number;
  pr?: number;
  pr_lng?: number;
  pr_yd?: number;
  pr_ypa?: number;
  pts_allow?: number;
  pts_allow_14_20?: number;
  pts_allow_21_27?: number;
  pts_allow_28_34?: number;
  pts_allow_35p?: number;
  pts_allow_7_13?: number;
  pts_half_ppr?: number;
  pts_idp?: number;
  pts_ppr?: number;
  pts_std?: number;
  punt_in_20?: number;
  punt_net_yd?: number;
  punt_tb?: number;
  punt_yds?: number;
  punts?: number;
  qb_hit?: number;
  rec?: number;
  rec_0_4?: number;
  rec_10_19?: number;
  rec_20_29?: number;
  rec_30_39?: number;
  rec_40p?: number;
  rec_5_9?: number;
  rec_air_yd?: number;
  rec_drop?: number;
  rec_fd?: number;
  rec_lng?: number;
  rec_rz_tgt?: number;
  rec_td?: number;
  rec_td_40p?: number;
  rec_td_50p?: number;
  rec_td_lng?: number;
  rec_tgt?: number;
  rec_yar?: number;
  rec_yd?: number;
  rec_ypr?: number;
  rec_ypt?: number;
  rush_40p?: number;
  rush_att?: number;
  rush_btkl?: number;
  rush_fd?: number;
  rush_lng?: number;
  rush_rec_yd?: number;
  rush_rz_att?: number;
  rush_td?: number;
  rush_td_40p?: number;
  rush_td_50p?: number;
  rush_td_lng?: number;
  rush_tkl_loss?: number;
  rush_tkl_loss_yd?: number;
  rush_yac?: number;
  rush_yd?: number;
  rush_ypa?: number;
  rz_att?: number;
  rz_conv?: number;
  rz_pct?: number;
  sack?: number;
  sack_yd?: number;
  st_ff?: number;
  st_fum_rec?: number;
  st_snp?: number;
  st_tkl_solo?: number;
  td?: number;
  tkl?: number;
  tkl_ast?: number;
  tkl_loss?: number;
  tkl_solo?: number;
  tm_def_snp?: number;
  tm_off_snp?: number;
  tm_st_snp?: number;
  xp_blkd?: number;
  xpa?: number;
  xpm?: number;
  xpmiss?: number;
  yds_allow?: number;
  yds_allow_100_199?: number;
  yds_allow_200_299?: number;
  yds_allow_300_349?: number;
  yds_allow_350_399?: number;
  yds_allow_400_449?: number;
  yds_allow_450_499: number | null | undefined
}